todos/requirements.txt
certifi==2018.1.18
chardet==3.0.4
dataclasses==0.4
Django==2.0.4
django-rest-auth==0.9.3
djangorestframework==3.7.7
djongo==1.2.23
idna==2.6
pymongo==3.6.1
pytz==2018.3
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
uWSGI==2.0.17